Added in route and request to delete a user's authtokens upon logging out

+++ src/client/views/auth/Logout.tsx
@@ -1,7 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
+import { api, User } from '../../utils/api-service';
 
 const Logout = () => {
-    localStorage.clear();
+
+    if (User && User.userid !== null && User.roles.includes('user')) {
+        const { userid } = User;
+        api(`/auth/logout/${userid}`);
+        localStorage.clear();
+    }
 
     return (
         <>
